+++ budget/web.py
@@ -34,17 +34,26 @@ main = Blueprint("main", __name__)
 mail = Mail()
 
 
-def requires_admin(f):
+def requires_admin(bypass=None):
     """Require admin permissions for @requires_admin decorated endpoints.
        Has no effect if ADMIN_PASSWORD is empty (default value)
+       The bypass variable is optionnal and used to conditionnaly bypass the admin authentication
+       It expects a tuple containing the name of an application setting and its expected value
+       e.g. if you use @require_admin(bypass=("ALLOW_PUBLIC_PROJECT_CREATION", True))
+            Admin authentication will be bypassed when ALLOW_PUBLIC_PROJECT_CREATION is set to True
     """
-    @wraps(f)
-    def admin_auth(*args, **kws):
-        is_admin = session.get('is_admin')
-        if is_admin or not current_app.config['ADMIN_PASSWORD']:
-            return f(*args, **kws)
-        raise Redirect303(url_for('.admin', goto=request.path))
-    return admin_auth
+    def check_admin(f):
+        @wraps(f)
+        def admin_auth(*args, **kws):
+            is_admin_auth_bypassed = False
+            if bypass is not None and current_app.config.get(bypass[0]) == bypass[1]:
+                is_admin_auth_bypassed = True
+            is_admin = session.get('is_admin')
+            if is_admin or is_admin_auth_bypassed:
+                return f(*args, **kws)
+            raise Redirect303(url_for('.admin', goto=request.path))
+        return admin_auth
+    return check_admin
 
 
 @main.url_defaults
@@ -157,18 +166,17 @@ def authenticate(project_id=None):
 def home():
     project_form = ProjectForm()
     auth_form = AuthenticationForm()
-    # If ADMIN_PASSWORD is empty we consider that admin mode is disabled
-    is_admin_mode_enabled = bool(current_app.config['ADMIN_PASSWORD'])
+    is_public_project_creation_allowed = current_app.config['ALLOW_PUBLIC_PROJECT_CREATION']
     is_demo_project_activated = current_app.config['ACTIVATE_DEMO_PROJECT']
 
     return render_template("home.html", project_form=project_form,
                            is_demo_project_activated=is_demo_project_activated,
-                           is_admin_mode_enabled=is_admin_mode_enabled,
+                           is_public_project_creation_allowed=is_public_project_creation_allowed,
                            auth_form=auth_form, session=session)
 
 
 @main.route("/create", methods=["GET", "POST"])
-@requires_admin
+@requires_admin(bypass=("ALLOW_PUBLIC_PROJECT_CREATION", True))
 def create_project():
     form = ProjectForm()
     if request.method == "GET" and 'project_id' in request.values: